Rename ChainerIO to PFIO
We rename the ChainerIO to PFIO as `Chainer` has entered the
maintenance phase.

An alias of `chainerio` has created to keep the backward compatibility.
