--- chainerio/chainer_extensions/__init__.py
@@ -1,1 +1,0 @@
-from chainerio.chainer_extensions.snapshot import load_snapshot  # NOQA
