+++ examples/seq2seq/seq2seq_chainerio.py
@@ -16,8 +16,8 @@ from chainer import training
 from chainer.training import extensions
 
 # CHAINERIO import
-import chainerio
-# chainerio.set_root("hdfs")
+import pfio
+# pfio.set_root("hdfs")
 # CHAINERIO import end
 
 
@@ -163,14 +163,14 @@ class CalculateBleu(chainer.training.Extension):
 
 def count_lines(path):
     # CHAINERIO add
-    with chainerio.open(path, mode='r') as f:
+    with pfio.open(path, mode='r') as f:
     # CHAINERIO add end
         return sum([1 for _ in f])
 
 
 def load_vocabulary(path):
     # CHAINERIO add
-    with chainerio.open(path, mode='r') as f:
+    with pfio.open(path, mode='r') as f:
     # CHAINERIO add end
         # +2 for UNK and EOS
         word_ids = {line.strip(): i + 2 for i, line in enumerate(f)}
@@ -186,7 +186,7 @@ def load_data(vocabulary, path):
     print('loading...: %s' % path)
 
     # CHAINERIO add
-    with chainerio.open(path, mode='r') as f:
+    with pfio.open(path, mode='r') as f:
     # CHAINERIO add end
             for line in bar(f, max_value=n_lines):
                 words = line.strip().split()
