+++ pfio/chainer_extensions/__init__.py
@@ -1,0 +1,1 @@
+from pfio.chainer_extensions.snapshot import load_snapshot  # NOQA
