+++ pfio/container.py
@@ -1,7 +1,7 @@
-from chainerio.io import IO
-from chainerio.profiler import IOProfiler
+from pfio.io import IO
+from pfio.profiler import IOProfiler
 
-from chainerio._typing import Optional
+from pfio._typing import Optional
 
 
 class Container(IO):
