+++ setup.py
@@ -18,18 +18,18 @@ for root, dirs, names in os.walk(templates):
 
 here = os.path.abspath(os.path.dirname(__file__))
 # Get __version__ variable
-exec(open(os.path.join(here, 'chainerio', 'version.py')).read())
+exec(open(os.path.join(here, 'pfio', 'version.py')).read())
 
 with open(os.path.join(here, 'README.md'), encoding='utf-8') as f:
     long_description = f.read()
 
 setup(
-    name='chainerio',
+    name='pfio',
     version=__version__,
-    description='Chainer IO library',
+    description='PFN IO library',
     author='Tianqi Xu, Kota Uenishi',
     author_email='tianqi@preferred.jp, kota@preferred.jp',
-    url='http://github.com/chainer/chainerio',
+    url='http://github.com/chainer/pfio',
     classifiers=[
         'Development Status :: 3 - Alpha',
 
@@ -49,7 +49,7 @@ setup(
     long_description=long_description,
     long_description_content_type='text/markdown',
     packages=find_packages(),
-    package_data={'chainerio': package_data},
+    package_data={'pfio': package_data},
     extras_require={'test': ['pytest', 'flake8', 'autopep8', 'parameterized'],
                     'doc': ['sphinx', 'sphinx_rtd_theme']},
     python_requires=">=3.5",
