+++ tox.ini
@@ -8,6 +8,6 @@ setenv =
         HOME = "/root"
 commands =
 	pytest tests -s -v
-	flake8 chainerio
+	flake8 chainerio pfio
 	flake8 tests
-	autopep8 -r chainerio tests --diff
+	autopep8 -r chainerio pfio tests --diff
